change got_pack_parse_offset_delta tslen argument to size_t
The tslen argument is always in the (1..9) range so `int' is technically
fine (even int8_t would be!) but all the callers are passing a size_t,
so change the type accordingly.

ok stsp@

+++ lib/got_lib_pack.h
@@ -41,7 +41,7 @@ const struct got_error *got_pack_start_privsep_child(struct got_pack *,
 const struct got_error *got_pack_close(struct got_pack *);
 
 const struct got_error *got_pack_parse_offset_delta(off_t *, size_t *,
-    struct got_pack *, off_t, int);
+    struct got_pack *, off_t, size_t);
 const struct got_error *got_pack_parse_ref_delta(struct got_object_id *,
     struct got_pack *, off_t, int);
 const struct got_error *got_pack_resolve_delta_chain(struct got_delta_chain *,
